server/requirements.txt
alabaster==0.7.12
asgiref==3.3.1
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Django==3.1.4
djangorestframework==3.12.2
docutils==0.16
idna==2.10
imagesize==1.2.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.4
packaging==20.8
pandas==1.1.5
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-igraph==0.8.3
pytz==2020.4
regex==2020.11.13
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
snowballstemmer==2.0.0
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
texttable==1.6.3
threadpoolctl==2.1.0
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
